+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import 'bulma/css/bulma.css';
 import './App.scss';
-import classNames from 'classnames';
+import { GoodsList } from './components/GoodsList';
 
 export const goods = [
   'Dumplings',
@@ -18,39 +18,6 @@ export const goods = [
 
 export const App = () => {
   const [selectProduct, setSelectProduct] = useState('Jam');
-
-  const handleSelectGoods = good => {
-    setSelectProduct(good === selectProduct ? null : good);
-  };
-
-  const GoodsList = ({ goodsList }) =>
-    goodsList.map(good => {
-      const isSelected = good === selectProduct; // Створено змінну для порівняння
-
-      return (
-        <tr
-          data-cy="Good"
-          className={classNames({
-            'has-background-success-light': isSelected,
-          })}
-        >
-          <td>
-            <button
-              data-cy={isSelected ? 'RemoveButton' : 'AddButton'}
-              type="button"
-              className={classNames('button', { 'is-info': isSelected })}
-              onClick={() => handleSelectGoods(good)}
-            >
-              {isSelected ? '-' : '+'}
-            </button>
-          </td>
-
-          <td data-cy="GoodTitle" className="is-vcentered">
-            {good}
-          </td>
-        </tr>
-      );
-    });
 
   return (
     <main className="section container">
@@ -72,7 +39,11 @@ export const App = () => {
 
       <table className="table">
         <tbody>
-          <GoodsList goodsList={goods} />
+          <GoodsList
+            goodsList={goods}
+            selectProduct={selectProduct}
+            setSelectProduct={setSelectProduct}
+          />
         </tbody>
       </table>
     </main>
